[Fix] Load action calls loop (#261)
* Fix: Remove faulty action calls from Savings

Because of savings-root.vue > <router-view> dependency on isStoreLoading -> isSavingsInfoLoading, during the Savings Info loading the root view displays alternative skeleton page then any error during Savings Info loading causes isSavingsInfoLoading flag to be flipped to the false and the fetch/load action is called. This causes the process to loop infinitely as the errored result is not cached / handler is not debounced in any way

* Fix: Remove faulty action calls from Smart Treasury

Because of treasury-root.vue > <router-view> dependency on isStoreLoading -> isTreasuryInfoLoading, during the Smart Treasury Info loading the root view displays alternative skeleton page then any error during Smart Treasury Info loading causes isTreasuryInfoLoading flag to be flipped to the false and the fetch/load action is called. This causes the process to loop infinitely as the errored result is not cached / handler is not debounced in any way